Modify run-jazzer.sh

+++ src/main/scala/utils/FileUtils.scala
@@ -7,11 +7,10 @@ object FileUtils {
 
   def getListOfFiles(dir: String):List[String] = {
     val d = new File(dir)
-    if (d.exists && d.isDirectory) {
-      d.listFiles.filter(_.isFile).map(_.toString).toList
-    } else {
-      List[String]()
-    }
+    if (!(d.exists && d.isDirectory))
+      throw new Exception(s"The directory ${d.getPath} does not exist or is not a directory")
+
+    d.listFiles.filter(_.isFile).map(_.toString).toList
   }
 
   def readFile(file: String): Seq[String] = {
@@ -22,7 +21,7 @@ object FileUtils {
   }
 
   def readDataset(path: String): Seq[String] = {
-    val file_parts = getListOfFiles(path).filter(path => path.startsWith("part-"))
+    val file_parts = getListOfFiles(path).filter(path => path.contains("part-"))
     file_parts.foldLeft(Seq[String]())((acc, e) => acc ++ readFile(e))
   }
 
@@ -33,9 +32,12 @@ object FileUtils {
 
   def writeToFile(data: Seq[String], path: String): Unit = {
     val file = new File(path)
-    file.getParentFile.mkdirs()
+
+    if (!file.getParentFile.exists() && !file.getParentFile.mkdirs())
+      throw new Exception("File could not be created")
+
     val bw = new BufferedWriter(new FileWriter(file))
-    data.foreach(r => bw.write(s"${r}\n"))
+    data.foreach(r => bw.write(s"$r\n"))
     bw.close()
   }
